Move the Base64 serializer out of the `base64` package
Move out the single file to remove the empty package.
