+++ ks3-standard/src/commonTest/kotlin/io/ks3/standard/ByteArrayAsBase64StringSerializerTest.kt
@@ -1,4 +1,4 @@
-package io.ks3.standard.base64
+package io.ks3.standard
 
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.FreeSpec
